waterAmount changed to float

+++ backend/plant_api/schemas.py
@@ -6,7 +6,7 @@ class UnitsSchema(Schema):
     name = fields.Str(metadata={"required": True}, validate=validate.Length(min=2, max=100))
     moistLimit = fields.Int(metadata={"required": True}, validate=validate.Range(min=0, max=100))
     waterTime = fields.Int(metadata={"required": True}, validate=validate.Range(min=0, max=600))
-    waterAmount = fields.Int(metadata={"required": True}, validate=validate.Range(min=0, max=600))
+    waterAmount = fields.Float(metadata={"required": True}, validate=validate.Range(min=0, max=600))
     enableAutoWatering = fields.Bool(metadata={"required": True})
     enableMaxWaterInterval = fields.Bool(metadata={"required": True})
     enableMinWaterInterval = fields.Bool(metadata={"required": True})
